Conflicts:
	mws/apimws/utils.py
	mws/sitesmanagement/models.py
	mws/sitesmanagement/views.py

+++ mws/sitesmanagement/views.py
@@ -1,3 +1,4 @@
+import bisect
 import datetime
 from django.contrib.auth.decorators import login_required
 from django.core.urlresolvers import reverse
@@ -13,7 +14,7 @@ from apimws.utils import email_confirmation, ip_register_api_request, launch_ans
 from mwsauth.utils import get_or_create_group_by_groupid, privileges_check
 from sitesmanagement.utils import is_camacuk, get_object_or_None
 from .models import SiteForm, DomainNameFormNew, BillingForm, DomainName, NetworkConfig, EmailConfirmation, \
-    VirtualMachine, SystemPackagesForm, Vhost, VhostForm, Site, UnixGroupForm, UnixGroup, SiteRequestDemo
+    VirtualMachine, Vhost, VhostForm, Site, UnixGroupForm, UnixGroup, SiteRequestDemo
 
 
 @login_required
@@ -419,7 +420,10 @@ def system_packages(request, vm_id):
     if vm.is_busy:
         return HttpResponseRedirect(reverse('sitesmanagement.views.show', kwargs={'site_id': site.id}))
 
-    ansible_configuraton = get_object_or_None(AnsibleConfiguration, vm=vm, key="System Packages")
+    ansible_configuraton = get_object_or_None(AnsibleConfiguration, vm=vm, key="system_packages")
+
+    packages_installed = list(int(x) for x in ansible_configuraton.value.split(",")) if ansible_configuraton is not \
+                                                                                         None else []
 
     breadcrumbs = {
         0: dict(name='Manage Web Service server: ' + str(site.name), url=reverse(show, kwargs={'site_id': site.id})),
@@ -428,29 +432,28 @@ def system_packages(request, vm_id):
         2: dict(name='System packages', url=reverse(system_packages, kwargs={'vm_id': vm.id}))
     }
 
+    package_number_list = [1,2,3,4] # TODO extract this to settings
+
     if request.method == 'POST':
-        system_packages_form = SystemPackagesForm(request.POST)
-        if system_packages_form.is_valid():
-            if ansible_configuraton is not None:
-                ansible_configuraton.value = ",".join(system_packages_form.cleaned_data.get('system_packages'))
-                ansible_configuraton.save()
+        package_number = int(request.POST['package_number'])
+        if package_number in package_number_list:
+            if packages_installed:
+                if package_number in packages_installed:
+                    packages_installed.remove(package_number)
+                    ansible_configuraton.value = ",".join(str(x) for x in packages_installed)
+                    ansible_configuraton.save()
+                else:
+                    bisect.insort_left(packages_installed, package_number)
+                    ansible_configuraton.value = ",".join(str(x) for x in packages_installed)
+                    ansible_configuraton.save()
             else:
-                AnsibleConfiguration.objects.create(vm=vm, key="System Packages",
-                                                    value=",".join(
-                                                        system_packages_form.cleaned_data.get('system_packages')))
+                AnsibleConfiguration.objects.create(vm=vm, key="system_packages",
+                                                    value=package_number)
             launch_ansible(vm)  # to install or delete new/old packages selected by the user
-            return HttpResponseRedirect(reverse('sitesmanagement.views.show',
-                                                kwargs={'site_id': site.id}))
-    else:
-        if ansible_configuraton is not None:
-            system_packages_form = SystemPackagesForm(initial={'system_packages': ansible_configuraton.
-                                                      value.split(",")})
-        else:
-            system_packages_form = SystemPackagesForm()
 
     return render(request, 'mws/system_packages.html', {
         'breadcrumbs': breadcrumbs,
-        'system_packages_form': system_packages_form,
+        'packages_installed': packages_installed,
         'vm': vm
     })
 
@@ -715,6 +718,23 @@ def add_domain(request, vhost_id, socket_error=None):
                         vhost.main_domain = new_domain
                         vhost.save()
                 launch_ansible(vhost.vm)  # to add the new domain name to the vhost apache configuration
+        else:
+            breadcrumbs = {
+                0: dict(name='Manage Web Service server: ' + str(site.name), url=reverse(show,
+                                                                                         kwargs={'site_id': site.id})),
+                1: dict(name='Server settings' if vhost.vm.primary else 'Test server settings',
+                        url=reverse(settings, kwargs={'vm_id': vhost.vm.id})),
+                2: dict(name='Web sites management: %s' % vhost.name, url=reverse(vhosts_management,
+                                                                               kwargs={'vm_id': vhost.vm.id})),
+                3: dict(name='Domain Names management', url=reverse(domains_management, kwargs={'vhost_id': vhost.id}))
+            }
+
+            return render(request, 'mws/domains.html', {
+                'breadcrumbs': breadcrumbs,
+                'vhost': vhost,
+                'domain_form': domain_form,
+                'error': True
+            })
 
     return redirect(reverse('sitesmanagement.views.domains_management', kwargs={'vhost_id': vhost.id}))
 
@@ -811,7 +831,11 @@ def delete_dn(request, domain_id):
         return HttpResponseRedirect(reverse('sitesmanagement.views.show', kwargs={'site_id': site.id}))
 
     if request.method == 'DELETE':
-        domain.delete()
+        if is_camacuk(domain.name):
+            domain.status = 'to_be_deleted'
+            domain.save()
+        else:
+            domain.delete()
         launch_ansible(vhost.vm)
         return HttpResponseRedirect(reverse('sitesmanagement.views.domains_management', kwargs={'vhost_id': vhost.id}))
